+++ ThirdPartyAdapters/bidmachine/CHANGELOG.md
@@ -1,5 +1,8 @@
 ## BidMachine Android Mediation Adapter Changelog
 
+#### Next Version
+- Maps GMA SDK underage flags to BidMachine SDK COPPA method.
+
 #### Version 3.5.1.0
 - Verified compatibility with BidMachine SDK version 3.5.1.